+++ mne_connectivity/io.py
@@ -1,3 +1,4 @@
+import numpy as np
 import xarray as xr
 
 from .base import (Connectivity, EpochConnectivity, EpochSpectralConnectivity,
@@ -57,11 +58,15 @@ def read_connectivity(fname):
         A connectivity class.
     """
     # open up a data-array using xarray
-    conn_da = xr.open_dataarray(fname)
+    # The engine specified requires the ability to save
+    # complex data types, which was not natively supported
+    # in xarray. Therefore, h5netcdf is the only engine
+    # to support that feature at this moment.
+    conn_da = xr.open_dataarray(fname, engine='h5netcdf')
 
     # map 'n/a' to 'None'
     for key, val in conn_da.attrs.items():
-        if not isinstance(val, list):
+        if not isinstance(val, list) and not isinstance(val, np.ndarray):
             if val == 'n/a':
                 conn_da.attrs[key] = None
     # get the name of the class